Move `config` module to `virtool_workflow`.
